chore(picker): demo拆解与规范
