+++ src/packages/picker/demo.tsx
@@ -1,396 +1,67 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { useTranslate } from '../../sites/assets/locale'
-import { Internation } from './doc.en'
-import Picker from './picker'
-import Cell from '@/packages/cell'
-import ConfigProvider from '@/packages/configprovider'
 import Demo1 from './demos/h5/demo1'
 import Demo2 from './demos/h5/demo2'
 import Demo3 from './demos/h5/demo3'
 import Demo4 from './demos/h5/demo4'
 import Demo5 from './demos/h5/demo5'
-
-interface T {
-  [props: string]: string
-}
-interface PickerOption {
-  text: string | number
-  value: string | number
-  disabled?: boolean
-  children?: PickerOption[]
-  className?: string | number
-}
+import Demo6 from './demos/h5/demo6'
+import Demo7 from './demos/h5/demo7'
+import Demo8 from './demos/h5/demo8'
 
 const PickerDemo = () => {
-  const [translated] = useTranslate<T>(Internation)
-
-  const [isVisible1, setIsVisible1] = useState(false)
-  const [isVisible2, setIsVisible2] = useState(false)
-  const [isVisible3, setIsVisible3] = useState(false)
-  const [isVisible4, setIsVisible4] = useState(false)
-  const [isVisible5, setIsVisible5] = useState(false)
-  const [isVisible6, setIsVisible6] = useState(false)
-  const [isVisible7, setIsVisible7] = useState(false)
-  const [isVisible8, setIsVisible8] = useState(false)
-
-  const [cityCustmer, setCityCustmer] = useState('')
-  const [baseDesc, setBaseDesc] = useState('')
-  const [baseDefault, setbaseDefault] = useState('')
-  const [mutilDesc, setMutilDesc] = useState('')
-  const [asyncDesc, setasyncDesc] = useState('')
-  const [tileDesc, settileDesc] = useState('')
-
-  const [defaultValue, setDefaultValue] = useState([2])
-
-  const listData1 = [
-    [
-      {
-        value: 1,
-        text: translated.nanJing,
-      },
-      {
-        value: 2,
-        text: translated.wuXi,
-      },
-      {
-        value: 3,
-        text: translated.zangZu,
-      },
-      {
-        value: 4,
-        text: translated.beiJing,
-      },
-      {
-        value: 5,
-        text: translated.lianYunGang,
-        disabled: true,
-      },
-      {
-        value: 6,
-        text: translated.wuhan,
-      },
-      {
-        value: 7,
-        text: translated.yangzhou,
-      },
-      {
-        value: 8,
-        text: translated.daQing,
-      },
-      {
-        value: 9,
-        text: translated.suiHua,
-      },
-      {
-        value: 10,
-        text: translated.weiFang,
-      },
-      {
-        value: 11,
-        text: translated.shijiazhuang,
-      },
-      {
-        value: 12,
-        text: translated.wlmq,
-      },
-    ],
-  ]
-
-  const listData2 = [
-    // 第一列
-    [
-      { text: translated.monday, value: 'Monday' },
-      { text: translated.tuesday, value: 'Tuesday' },
-      { text: translated.wednesday, value: 'Wednesday' },
-      { text: translated.thursday, value: 'Thursday' },
-      { text: translated.friday, value: 'Friday' },
-    ],
-    // 第二列
-    [
-      { text: translated.morning, value: 'Morning' },
-      { text: translated.afternoon, value: 'Afternoon' },
-      { text: translated.evening, value: 'Evening' },
-    ],
-  ]
-
-  const custmerCityData = [
-    {
-      value: 1,
-      text: translated.beiJing,
-      children: [
-        {
-          value: 1,
-          text: translated.chaoYang,
-        },
-        {
-          value: 2,
-          text: translated.haiDian,
-        },
-        {
-          value: 3,
-          text: translated.daXing,
-        },
-        {
-          value: 4,
-          text: translated.dongCheng,
-        },
-        {
-          value: 5,
-          text: translated.xiCheng,
-        },
-        {
-          value: 6,
-          text: translated.fengTai,
-        },
-      ],
+  const [translated] = useTranslate({
+    'zh-CN': {
+      basic: '基础用法',
+      defaultSelected: '默认选中项',
+      controlled: '受控',
+      tileDesc: '平铺展示',
+      multipleColumns: '多列样式',
+      cascade: '多级联动',
+      async: '异步获取',
+      theme: '自定义主题',
     },
-    {
-      value: 2,
-      text: translated.shangHai,
-      children: [
-        {
-          value: 1,
-          text: translated.huangPu,
-        },
-        {
-          value: 2,
-          text: translated.changNing,
-        },
-        {
-          value: 3,
-          text: translated.puTuo,
-        },
-        {
-          value: 4,
-          text: translated.yangPu,
-        },
-        {
-          value: 5,
-          text: translated.puDong,
-        },
-      ],
+    'zh-TW': {
+      basic: '基础用法',
+      defaultSelected: '默認選中項',
+      controlled: '受控',
+      tileDesc: '平鋪展示',
+      multipleColumns: '多列樣式',
+      cascade: '多級聯動',
+      async: '異步獲取',
+      theme: '自定義主题',
     },
-  ]
-
-  // 动态获取
-  const [asyncData, setAsyncData] = useState([
-    {
-      value: 1,
-      text: translated.beiJing,
-      children: [
-        {
-          value: 1,
-          text: translated.chaoYang,
-        },
-        {
-          value: 2,
-          text: translated.haiDian,
-        },
-        {
-          value: 3,
-          text: translated.daXing,
-        },
-        {
-          value: 4,
-          text: translated.dongCheng,
-        },
-        {
-          value: 5,
-          text: translated.xiCheng,
-        },
-        {
-          value: 6,
-          text: translated.fengTai,
-        },
-      ],
+    'en-US': {
+      basic: 'Basic Usage',
+      defaultSelected: 'Default Index',
+      controlled: 'Controlled',
+      tileDesc: 'Tile',
+      multipleColumns: 'Multiple Columns',
+      cascade: 'Cascade',
+      async: 'Async',
+      theme: 'Custom Theme',
     },
-    {
-      value: 2,
-      text: translated.shangHai,
-      children: [],
-    },
-  ])
+  })
 
-  const setChooseValueCustmer = (
-    options: PickerOption[],
-    values: (string | number)[]
-  ) => {
-    const str = options.map((item) => item.text).join('-')
-    setCityCustmer(str)
-  }
-
-  const setAsyncConfirm = (
-    options: PickerOption[],
-    values: (string | number)[]
-  ) => {
-    console.log(values, options)
-    const str = options.map((item) => item.text).join('-')
-    setasyncDesc(str)
-  }
-
-  const updateChooseValueCustmer = (
-    options: PickerOption[],
-    values: (string | number)[],
-    columnIndex: number
-  ) => {
-    console.log('updateChooseValueCustmer', columnIndex, values, options)
-    if (columnIndex === 0 && values[0] === 2) {
-      setTimeout(() => {
-        if (asyncData[1].children.length === 0) {
-          asyncData[1].children = [
-            {
-              value: 1,
-              text: translated.huangPu,
-            },
-            {
-              value: 2,
-              text: translated.changNing,
-            },
-            {
-              value: 3,
-              text: translated.puTuo,
-            },
-            {
-              value: 4,
-              text: translated.yangPu,
-            },
-            {
-              value: 5,
-              text: translated.puDong,
-            },
-          ]
-          setAsyncData([...asyncData])
-        }
-      }, 100)
-    }
-  }
-
-  const [val, setVal] = useState<Array<number | string>>([])
-  // 确定选择
-  const confirmPicker = (
-    type: string,
-    options: PickerOption[],
-    values: (string | number)[]
-  ) => {
-    console.log('demo 确定', options, values)
-    let description = ''
-    options.forEach((option: any) => {
-      description += ` ${option.text}`
-    })
-    if (type === 'base') {
-      setBaseDesc(description)
-      setIsVisible7(false)
-    }
-    if (type === 'mutil') {
-      setMutilDesc(description)
-      setIsVisible2(false)
-    }
-
-    if (type === 'default') {
-      setbaseDefault(description)
-      setDefaultValue([options[0].value as number])
-      setIsVisible4(false)
-    }
-
-    if (type === 'tile') {
-      settileDesc(description)
-      setIsVisible6(false)
-    }
-  }
   return (
     <>
       <div className="demo">
         <h2>{translated.basic}</h2>
         <Demo1 />
-
         <h2>{translated.defaultSelected}</h2>
         <Demo2 />
-
         <h2>{translated.controlled}</h2>
         <Demo3 />
-
         <h2>{translated.multipleColumns}</h2>
         <Demo4 />
         <h2>{translated.tileDesc}</h2>
         <Demo5 />
-
         <h2>{translated.cascade}</h2>
-        <Cell
-          title={translated.cascade}
-          description={cityCustmer}
-          onClick={() => setIsVisible3(!isVisible3)}
-        />
-
-        <Picker
-          visible={isVisible3}
-          options={custmerCityData}
-          onClose={() => setIsVisible3(false)}
-          onConfirm={(list, values) => setChooseValueCustmer(list, values)}
-          onChange={(
-            options: PickerOption[],
-            value: (string | number)[],
-            columnIndex: number
-          ) =>
-            console.log(
-              asyncData,
-              translated.cascade,
-              columnIndex,
-              value,
-              options
-            )
-          }
-        />
-
+        <Demo6 />
         <h2>{translated.async}</h2>
-        <Cell
-          title={translated.chooseCity}
-          description={asyncDesc}
-          onClick={() => setIsVisible5(!isVisible5)}
-        />
-
-        <Picker
-          visible={isVisible5}
-          options={asyncData}
-          onClose={() => setIsVisible5(false)}
-          onConfirm={(list, values) => setAsyncConfirm(list, values)}
-          onChange={(
-            selectedOptions: PickerOption[],
-            selectedValue: (string | number)[],
-            columnIndex: number
-          ) =>
-            updateChooseValueCustmer(
-              selectedOptions,
-              selectedValue,
-              columnIndex
-            )
-          }
-        />
-
+        <Demo7 />
         <h2>{translated.theme}</h2>
-        <Cell
-          title={translated.chooseCity}
-          description={asyncDesc}
-          onClick={() => setIsVisible8(!isVisible8)}
-        />
-        <ConfigProvider
-          theme={{
-            nutuiPickerItemHeight: '48px',
-            nutuiPickerItemActiveLineBorder:
-              '1px dashed var(--nutui-color-primary)',
-            nutuiPickerItemTextColor: 'var(--nutui-color-primary)',
-          }}
-        >
-          <Picker
-            title={translated.chooseCity}
-            visible={isVisible8}
-            options={listData1}
-            onConfirm={(list, values) => confirmPicker('base', list, values)}
-            onClose={() => {
-              setIsVisible8(false)
-              console.log('onclose')
-            }}
-          />
-        </ConfigProvider>
+        <Demo8 />
       </div>
     </>
   )
