+++ src/packages/picker/demos/taro/demo2.tsx
@@ -1,10 +1,18 @@
 import React, { useState } from 'react'
 import { Picker, Cell } from '@nutui/nutui-react-taro'
 
+interface PickerOption {
+  text: string | number
+  value: string | number
+  disabled?: boolean
+  children?: PickerOption[]
+  className?: string | number
+}
 const Demo2 = () => {
   const [visible, setVisible] = useState(false)
   const [baseDefault, setbaseDefault] = useState('')
-  const [defaultValue, setDefaultValue] = useState([2])
+  const [defaultValue] = useState([2])
+
   const listData1 = [
     [
       { value: 1, text: '南京市' },
@@ -24,7 +32,7 @@ const Demo2 = () => {
   ) => {
     let description = ''
     options.forEach((option: any) => {
-      description += option.text
+      description += ` ${option.text}`
     })
     setbaseDefault(description)
   }
@@ -36,6 +44,7 @@ const Demo2 = () => {
         onClick={() => setVisible(!visible)}
       />
       <Picker
+        title="请选择城市"
         visible={visible}
         options={listData1}
         defaultValue={defaultValue}
