+++ src/packages/picker/demos/taro/demo3.tsx
@@ -1,23 +1,28 @@
 import React, { useState } from 'react'
 import { Picker, Cell } from '@nutui/nutui-react-taro'
 
+interface PickerOption {
+  text: string | number
+  value: string | number
+  disabled?: boolean
+  children?: PickerOption[]
+  className?: string | number
+}
 const Demo3 = () => {
-  const [isVisible2, setIsVisible2] = useState(false)
-  const [mutilDesc, setMutilDesc] = useState('')
-  const listData2 = [
-    // 第一列
+  const [isVisible, setIsVisible] = useState(false)
+  const [baseDesc, setBaseDesc] = useState('')
+  const [val, setVal] = useState<Array<number | string>>([])
+  const options = [
     [
-      { text: '周一', value: 'Monday' },
-      { text: '周二', value: 'Tuesday' },
-      { text: '周三', value: 'Wednesday' },
-      { text: '周四', value: 'Thursday' },
-      { text: '周五', value: 'Friday' },
-    ],
-    // 第二列
-    [
-      { text: '上午', value: 'Morning' },
-      { text: '下午', value: 'Afternoon' },
-      { text: '晚上', value: 'Evening' },
+      { value: 1, text: '南京市' },
+      { value: 2, text: '无锡市' },
+      { value: 3, text: '海北藏族自治区' },
+      { value: 4, text: '北京市' },
+      { value: 5, text: '连云港市' },
+      { value: 8, text: '大庆市' },
+      { value: 9, text: '绥化市' },
+      { value: 10, text: '潍坊市' },
+      { value: 12, text: '乌鲁木齐市' },
     ],
   ]
   const confirmPicker = (
@@ -26,23 +31,29 @@ const Demo3 = () => {
   ) => {
     let description = ''
     options.forEach((option: any) => {
-      description += option.text
+      description += ` ${option.text}`
     })
-    setbaseDefault(description)
+    setBaseDesc(description)
   }
   return (
     <>
       <Cell
-        title="多列用法"
-        description={mutilDesc}
-        onClick={() => setIsVisible2(!isVisible2)}
+        title="请选择城市"
+        description={baseDesc}
+        onClick={() => setIsVisible(!isVisible)}
       />
       <Picker
-        visible={isVisible2}
-        options={listData2}
-        onClose={() => setIsVisible2(false)}
-        defaultValue={['Wednesday']}
-        onConfirm={(list, values) => confirmPicker(list, values)}
+        title="请选择城市"
+        visible={isVisible}
+        value={val}
+        options={options}
+        onConfirm={(list, values) => {
+          confirmPicker(list, values)
+          setVal(values)
+        }}
+        onClose={() => {
+          setIsVisible(false)
+        }}
       />
     </>
   )
