+++ src/packages/picker/demos/taro/demo4.tsx
@@ -1,20 +1,30 @@
 import React, { useState } from 'react'
 import { Picker, Cell } from '@nutui/nutui-react-taro'
 
+interface PickerOption {
+  text: string | number
+  value: string | number
+  disabled?: boolean
+  children?: PickerOption[]
+  className?: string | number
+}
 const Demo4 = () => {
-  const [visible, setIsVisible] = useState(false)
-  const [tileDesc, settileDesc] = useState('')
-  const options = [
+  const [isVisible2, setIsVisible2] = useState(false)
+  const [mutilDesc, setMutilDesc] = useState('')
+  const listData2 = [
+    // 第一列
     [
-      { value: 1, text: '南京市' },
-      { value: 2, text: '无锡市' },
-      { value: 3, text: '海北藏族自治区' },
-      { value: 4, text: '北京市' },
-      { value: 5, text: '连云港市' },
-      { value: 8, text: '大庆市' },
-      { value: 9, text: '绥化市' },
-      { value: 10, text: '潍坊市' },
-      { value: 12, text: '乌鲁木齐市' },
+      { text: '周一', value: 'Monday' },
+      { text: '周二', value: 'Tuesday' },
+      { text: '周三', value: 'Wednesday' },
+      { text: '周四', value: 'Thursday' },
+      { text: '周五', value: 'Friday' },
+    ],
+    // 第二列
+    [
+      { text: '上午', value: 'Morning' },
+      { text: '下午', value: 'Afternoon' },
+      { text: '晚上', value: 'Evening' },
     ],
   ]
   const confirmPicker = (
@@ -23,24 +33,27 @@ const Demo4 = () => {
   ) => {
     let description = ''
     options.forEach((option: any) => {
-      description += option.text
+      description += ` ${option.text}`
     })
-    settileDesc(description)
+    setMutilDesc(description)
+  }
+  const changePicker = (options: any[], values: any, columnIndex: number) => {
+    console.log('picker onChange', columnIndex, values, options)
   }
   return (
     <>
       <Cell
-        title="请选择城市"
-        description={settileDesc}
-        onClick={() => setIsVisible(!visible)}
+        title="多列用法"
+        description={mutilDesc}
+        onClick={() => setIsVisible2(!isVisible2)}
       />
       <Picker
-        visible={visible}
-        options={options}
-        threeDimensional={false}
-        duration={1000}
+        visible={isVisible2}
+        options={listData2}
+        onClose={() => setIsVisible2(false)}
+        defaultValue={['Wednesday']}
+        onChange={changePicker}
         onConfirm={(list, values) => confirmPicker(list, values)}
-        onClose={() => setIsVisible(false)}
       />
     </>
   )
