--- src/packages/picker/doc.en.ts
@@ -1,145 +1,0 @@
-export const Internation = {
-  'zh-CN': {
-    basic: '基础用法',
-    chooseCity: '请选择城市',
-    defaultSelected: '默认选中项',
-    controlled: '受控',
-    tileDesc: '平铺展示',
-    multipleColumns: '多列样式',
-    chooseTime: '请选择时间',
-    cascade: '多级联动',
-    async: '异步获取',
-    theme: '主题',
-    validTime: '请选择有效日期',
-    chooseDate: '选择日期',
-    always: '永远有效',
-    monday: '周一',
-    tuesday: '周二',
-    wednesday: '周三',
-    thursday: '周四',
-    friday: '周五',
-    morning: '上午',
-    afternoon: '下午',
-    evening: '晚上',
-    nanJing: '南京市',
-    wuXi: '无锡市',
-    zangZu: '海北藏族自治区',
-    beiJing: '北京市',
-    lianYunGang: '连云港市',
-    wuhan: '武汉市',
-    yangzhou: '扬州市',
-    daQing: '大庆市',
-    suiHua: '绥化市',
-    weiFang: '潍坊市',
-    shijiazhuang: '石家庄市',
-    wlmq: '乌鲁木齐市',
-    hangZhou: '杭州市',
-    chaoYang: '朝阳区',
-    haiDian: '海淀区',
-    daXing: '大兴区',
-    dongCheng: '东城区',
-    xiCheng: '西城区',
-    fengTai: '丰台区',
-    shangHai: '上海市',
-    huangPu: '黄浦区',
-    changNing: '长宁区',
-    puTuo: '普陀区',
-    yangPu: '杨浦区',
-    puDong: '浦东新区',
-  },
-  'zh-TW': {
-    basic: '基础用法',
-    chooseCity: '請選擇城市',
-    defaultSelected: '默認選中項',
-    controlled: '受控',
-    tileDesc: '平鋪展示',
-    multipleColumns: '多列樣式',
-    chooseTime: '請選擇時間',
-    cascade: '多級聯動',
-    async: '異步獲取',
-    theme: '主题',
-    chooseDate: '選擇日期',
-    always: '永遠有效',
-    monday: '周一',
-    tuesday: '周二',
-    wednesday: '周三',
-    thursday: '周四',
-    friday: '周五',
-    morning: '上午',
-    afternoon: '下午',
-    evening: '晚上',
-    nanJing: '南京市',
-    wuXi: '無錫市',
-    zangZu: '海北藏族自治區',
-    beiJing: '北京市',
-    lianYunGang: '連雲港市',
-    wuhan: '武漢市',
-    yangzhou: '江蘇市',
-    daQing: '大慶市',
-    suiHua: '綏化市',
-    weiFang: '潍坊市',
-    shijiazhuang: '石家莊市',
-    wlmq: '烏魯木齊市',
-    hangZhou: '杭州市',
-    chaoYang: '朝陽區',
-    haiDian: '海澱區',
-    daXing: '大興區',
-    dongCheng: '東城區',
-    xiCheng: '西城區',
-    fengTai: '豐臺區',
-    shangHai: '上海市',
-    huangPu: '黃浦區',
-    changNing: '長寧區',
-    puTuo: '普陀區',
-    yangPu: '楊浦區',
-    puDong: '浦東新區',
-  },
-  'en-US': {
-    basic: 'Basic Usage',
-    chooseCity: 'Choose City',
-    defaultSelected: 'Default Index',
-    controlled: 'Controlled',
-    tileDesc: 'Tile',
-    multipleColumns: 'Multiple Columns',
-    chooseTime: 'Choose Time',
-    cascade: 'Cascade',
-    async: 'Async',
-    theme: 'Theme',
-    validTime: 'Valid Time',
-    chooseDate: 'Choose Date',
-    always: 'Always',
-    monday: 'Monday',
-    tuesday: 'Tuesday',
-    wednesday: 'Wednesday',
-    thursday: 'Thursday',
-    friday: 'Friday',
-    morning: 'Morning',
-    afternoon: 'Afternoon',
-    evening: 'Evening',
-    nanJing: 'NanJing',
-    wuXi: 'WuXi',
-    zangZu: 'ZangZu',
-    beiJing: 'BeiJing',
-    lianYunGang: 'LianYunGang',
-    wuhan: 'WuHan',
-    yangzhou: 'JiangSu',
-    daQing: 'DaQing',
-    suiHua: 'SuiHua',
-    weiFang: 'WeiFang',
-    shijiazhuang: 'ShiJiaZhuang',
-    wlmq: 'Urumqi Municipality',
-    hangZhou: 'HangZhou',
-    chaoYang: 'ChaoYang',
-    haiDian: 'HaiDian',
-    daXing: 'DaXing',
-    dongCheng: 'DongCheng',
-    xiCheng: 'XiCheng',
-    fengTai: 'FengTai',
-    shangHai: 'ShangHai',
-    huangPu: 'HuangPu',
-    changNing: 'ChangNing',
-    puTuo: 'PuTuo',
-    yangPu: 'YangPu',
-    puDong: 'PuDong',
-  },
-}
